requirements.txt
alembic==1.5.4
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Migrate==2.6.0
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
Mako==1.1.4
MarkupSafe==1.1.1
matplotlib==3.2.0
numpy==1.19.4
pandas==1.1.5
Pillow==8.0.1
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.4
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.2
Werkzeug==1.0.1
